Refactor link commands
