+++ commands/unlink.ts
@@ -4,8 +4,8 @@ import {
   SlashCommandBuilder,
 } from 'discord.js';
 import { createCommand } from '../types/DiscordCommand';
-import { IChannelLink } from '../db/models/ChannelLink';
 import cache from '../cache';
+import { unlinkChannel } from './utilities/unlinking';
 
 const UnlinkOptions = {
   SOURCE_CHANNEL: 'source-channel',
@@ -29,28 +29,6 @@ const data = new SlashCommandBuilder()
       )
   )
   .setDefaultMemberPermissions(PermissionFlagsBits.ManageGuild);
-
-export const unlinkChannel = async (
-  channelLink: IChannelLink,
-  channelId: string
-) => {
-  if (channelLink.links.find((link) => link.id === channelId)) {
-    channelLink.links = channelLink.links.filter(
-      (trChannel) => trChannel.id !== channelId
-    );
-
-    if (channelLink.links.length === 0) {
-      // remove channelLink since it doesn't have any links anymore
-      // because it cannot be called "channelLink" with no links
-      await cache.channelLink.delete(channelLink.id);
-    } else {
-      await cache.channelLink.update(channelLink);
-    }
-
-    return true;
-  }
-  return false;
-};
 
 const execute = async (interaction: ChatInputCommandInteraction) => {
   const sourceChannelId =
